On-policy CRAM Bandit

This function performs simulation for cram bandit policy learning and evaluation.

Usage

cram_bandit_sim(
  horizon,
  simulations,
  bandit,
  policy,
  alpha = 0.05,
  do_parallel = FALSE,
  seed = 42
)

Arguments

horizon

The number of timesteps

simulations

The number of simulations

bandit

The bandit, generating contextual features and observed rewards

policy

The policy, choosing the arm at each timestep

alpha

Significance level for confidence intervals for calculating the empirical coverage. Default is 0.05 (95% confidence).

do_parallel

Whether to parallelize the simulations. Default to FALSE.

seed

Seed for simulation

Value

A **list** containing:

raw_results

A data frame summarizing key metrics: Average Prediction Error on the Policy Value Estimate, Average Prediction Error on the Variance Estimate, Empirical Coverage of the alpha level Confidence Interval.

interactive_table

An interactive table summarizing key metrics for detailed exploration.