Set Model

This function maps the model type and learner type to the corresponding model function.

Usage

set_model(model_type, learner_type, model_params)

Arguments

model_type

The model type for policy learning. Options include "causal_forest", "s_learner", and "m_learner". Default is "causal_forest".

learner_type

The learner type for the chosen model. Options include "ridge" for Ridge Regression and "fnn" for Feedforward Neural Network. Default is "ridge".

model_params

A list of additional parameters to pass to the model, which can be any parameter defined in the model reference package. Defaults to NULL. For FNNs, the following elements are defined in the model params list:

input_layer

A list defining the input layer. Must include:

units

Number of units in the input layer.

activation

Activation function for the input layer.

input_shape

Input shape for the layer.

layers

A list of lists, where each sublist specifies a hidden layer with:

units

Number of units in the layer.

activation

Activation function for the layer.

output_layer

A list defining the output layer. Must include:

units

Number of units in the output layer.

activation

Activation function for the output layer (e.g., "linear" or "sigmoid").

compile_args

A list of arguments for compiling the model. Must include:

optimizer

Optimizer for training (e.g., "adam" or "sgd").

loss

Loss function (e.g., "mse" or "binary_crossentropy").

metrics

Optional list of metrics for evaluation (e.g., c("accuracy")).

For other learners (e.g., "ridge" or "causal_forest"), model_params can include relevant hyperparameters.

Value

The instantiated model object or the corresponding model function.

Examples

# Example: Causal Forest with default parameters
set_model("causal_forest", NULL, model_params = list(num.trees = 100))
#> function (X, Y, W, Y.hat = NULL, W.hat = NULL, num.trees = 2000, 
#>     sample.weights = NULL, clusters = NULL, equalize.cluster.weights = FALSE, 
#>     sample.fraction = 0.5, mtry = min(ceiling(sqrt(ncol(X)) + 
#>         20), ncol(X)), min.node.size = 5, honesty = TRUE, honesty.fraction = 0.5, 
#>     honesty.prune.leaves = TRUE, alpha = 0.05, imbalance.penalty = 0, 
#>     stabilize.splits = TRUE, ci.group.size = 2, tune.parameters = "none", 
#>     tune.num.trees = 200, tune.num.reps = 50, tune.num.draws = 1000, 
#>     compute.oob.predictions = TRUE, num.threads = NULL, seed = runif(1, 
#>         0, .Machine$integer.max)) 
#> {
#>     has.missing.values <- validate_X(X, allow.na = TRUE)
#>     validate_sample_weights(sample.weights, X)
#>     Y <- validate_observations(Y, X)
#>     W <- validate_observations(W, X)
#>     clusters <- validate_clusters(clusters, X)
#>     samples.per.cluster <- validate_equalize_cluster_weights(equalize.cluster.weights, 
#>         clusters, sample.weights)
#>     num.threads <- validate_num_threads(num.threads)
#>     all.tunable.params <- c("sample.fraction", "mtry", "min.node.size", 
#>         "honesty.fraction", "honesty.prune.leaves", "alpha", 
#>         "imbalance.penalty")
#>     default.parameters <- list(sample.fraction = 0.5, mtry = min(ceiling(sqrt(ncol(X)) + 
#>         20), ncol(X)), min.node.size = 5, honesty.fraction = 0.5, 
#>         honesty.prune.leaves = TRUE, alpha = 0.05, imbalance.penalty = 0)
#>     args.orthog <- list(X = X, num.trees = max(50, num.trees/4), 
#>         sample.weights = sample.weights, clusters = clusters, 
#>         equalize.cluster.weights = equalize.cluster.weights, 
#>         sample.fraction = sample.fraction, mtry = mtry, min.node.size = 5, 
#>         honesty = TRUE, honesty.fraction = 0.5, honesty.prune.leaves = honesty.prune.leaves, 
#>         alpha = alpha, imbalance.penalty = imbalance.penalty, 
#>         ci.group.size = 1, tune.parameters = tune.parameters, 
#>         num.threads = num.threads, seed = seed)
#>     if (is.null(Y.hat)) {
#>         forest.Y <- do.call(regression_forest, c(Y = list(Y), 
#>             args.orthog))
#>         Y.hat <- predict(forest.Y)$predictions
#>     }
#>     else if (length(Y.hat) == 1) {
#>         Y.hat <- rep(Y.hat, nrow(X))
#>     }
#>     else if (length(Y.hat) != nrow(X)) {
#>         stop("Y.hat has incorrect length.")
#>     }
#>     if (is.null(W.hat)) {
#>         forest.W <- do.call(regression_forest, c(Y = list(W), 
#>             args.orthog))
#>         W.hat <- predict(forest.W)$predictions
#>     }
#>     else if (length(W.hat) == 1) {
#>         W.hat <- rep(W.hat, nrow(X))
#>     }
#>     else if (length(W.hat) != nrow(X)) {
#>         stop("W.hat has incorrect length.")
#>     }
#>     Y.centered <- Y - Y.hat
#>     W.centered <- W - W.hat
#>     data <- create_train_matrices(X, outcome = Y.centered, treatment = W.centered, 
#>         sample.weights = sample.weights)
#>     args <- list(num.trees = num.trees, clusters = clusters, 
#>         samples.per.cluster = samples.per.cluster, sample.fraction = sample.fraction, 
#>         mtry = mtry, min.node.size = min.node.size, honesty = honesty, 
#>         honesty.fraction = honesty.fraction, honesty.prune.leaves = honesty.prune.leaves, 
#>         alpha = alpha, imbalance.penalty = imbalance.penalty, 
#>         stabilize.splits = stabilize.splits, ci.group.size = ci.group.size, 
#>         compute.oob.predictions = compute.oob.predictions, num.threads = num.threads, 
#>         seed = seed, reduced.form.weight = 0, legacy.seed = get_legacy_seed())
#>     tuning.output <- NULL
#>     if (!identical(tune.parameters, "none")) {
#>         if (identical(tune.parameters, "all")) {
#>             tune.parameters <- all.tunable.params
#>         }
#>         else {
#>             tune.parameters <- unique(match.arg(tune.parameters, 
#>                 all.tunable.params, several.ok = TRUE))
#>         }
#>         if (!honesty) {
#>             tune.parameters <- tune.parameters[!grepl("honesty", 
#>                 tune.parameters)]
#>         }
#>         tune.parameters.defaults <- default.parameters[tune.parameters]
#>         tuning.output <- tune_forest(data = data, nrow.X = nrow(X), 
#>             ncol.X = ncol(X), args = args, tune.parameters = tune.parameters, 
#>             tune.parameters.defaults = tune.parameters.defaults, 
#>             tune.num.trees = tune.num.trees, tune.num.reps = tune.num.reps, 
#>             tune.num.draws = tune.num.draws, train = causal_train)
#>         args <- utils::modifyList(args, as.list(tuning.output[["params"]]))
#>     }
#>     forest <- do.call.rcpp(causal_train, c(data, args))
#>     class(forest) <- c("causal_forest", "grf")
#>     forest[["seed"]] <- seed
#>     forest[["num.threads"]] <- num.threads
#>     forest[["ci.group.size"]] <- ci.group.size
#>     forest[["X.orig"]] <- X
#>     forest[["Y.orig"]] <- Y
#>     forest[["W.orig"]] <- W
#>     forest[["Y.hat"]] <- Y.hat
#>     forest[["W.hat"]] <- W.hat
#>     forest[["clusters"]] <- clusters
#>     forest[["equalize.cluster.weights"]] <- equalize.cluster.weights
#>     forest[["sample.weights"]] <- sample.weights
#>     forest[["tunable.params"]] <- args[all.tunable.params]
#>     forest[["tuning.output"]] <- tuning.output
#>     forest[["has.missing.values"]] <- has.missing.values
#>     forest
#> }
#> <bytecode: 0x000001d520133268>
#> <environment: namespace:grf>