Validate or Set the Baseline Policy

This function validates a provided baseline policy or sets a default baseline policy of zeros for all individuals.

Usage

test_baseline_policy(baseline_policy, n)

Arguments

baseline_policy

A list representing the baseline policy for each individual. If NULL, a default baseline policy of zeros is created.

n

An integer specifying the number of individuals in the population.

Value

A validated or default baseline policy as a list of numeric values.

Examples

# Example: Default baseline policy
baseline_policy <- test_baseline_policy(NULL, n = 10)

# Example: Valid baseline policy
valid_policy <- as.list(rep(1, 10))
baseline_policy <- test_baseline_policy(valid_policy, n = 10)

# Example: Invalid baseline policy
if (FALSE) { # \dontrun{
invalid_policy <- c(1, 0, 1, 0)
baseline_policy <- test_baseline_policy(invalid_policy, n = 10)
} # }